+++ build.sbt
@@ -2,7 +2,7 @@ name := "ClickThroughRate"
 
 version := "1.0"
 
-scalaVersion := "2.10.4"
+scalaVersion := "2.10.5"
 
 assemblyJarName in assembly := "CTR.jar"
