--- src/main/scala/main.scala
@@ -1,5 +1,0 @@
-object Main{
-  def main(args:Array[String]): Unit ={
-    print("Initial")
-  }
-}
