--- src/main/scala/mllib/randomForest/RandomForest.scala
@@ -1,12 +1,0 @@
-package mllib.randomForest
-
-import mllib.SmallModel
-
-/**
-  * Created by benjamin658 on 2016/5/26.
-  */
-class RandomForest extends SmallModel{
-  def train(): Unit ={
-
-  }
-}
